pyproject.toml
[tool.poetry]
name = "pythonproject11"
version = "0.1.0"
description = ""
authors = ["Nastya Foxe <Kukusiki2204@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.8.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
